refactor: классы обработчиков команд вынесены в отедльный пакет и изменено название интерфейса CommandHandler
