+++ src/main/java/com/github/mikesafonov/operatorbot/handler/MessageHandler.java
@@ -4,6 +4,6 @@ import com.github.mikesafonov.operatorbot.command.ParsedCommand;
 import com.github.mikesafonov.operatorbot.service.AuthorizationTelegram;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 
-public interface CommandHandler {
+public interface MessageHandler {
     SendMessage operate(long chatId, AuthorizationTelegram user, ParsedCommand parsedCommand);
 }
