+++ src/main/java/com/github/mikesafonov/operatorbot/handler/command/DefaultHandler.java
@@ -1,10 +1,11 @@
-package com.github.mikesafonov.operatorbot.handler;
+package com.github.mikesafonov.operatorbot.handler.command;
 
 import com.github.mikesafonov.operatorbot.command.ParsedCommand;
+import com.github.mikesafonov.operatorbot.handler.MessageHandler;
 import com.github.mikesafonov.operatorbot.service.AuthorizationTelegram;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 
-public class DefaultHandler implements CommandHandler {
+public class DefaultHandler implements MessageHandler {
     @Override
     public SendMessage operate(long chatId, AuthorizationTelegram user, ParsedCommand parsedCommand) {
         return getMessageDefault(chatId);
