+++ src/test/java/com/github/mikesafonov/operatorbot/messageReciever/MessageHandlerFactoryTest.java
@@ -3,8 +3,9 @@ package com.github.mikesafonov.operatorbot.messageReciever;
 import com.github.mikesafonov.operatorbot.command.Command;
 import com.github.mikesafonov.operatorbot.command.Parser;
 import com.github.mikesafonov.operatorbot.handler.*;
-import com.github.mikesafonov.operatorbot.handler.admin.AddHandler;
-import com.github.mikesafonov.operatorbot.handler.internal.WhoHandler;
+import com.github.mikesafonov.operatorbot.handler.command.admin.AddHandler;
+import com.github.mikesafonov.operatorbot.handler.command.*;
+import com.github.mikesafonov.operatorbot.handler.command.internal.WhoHandler;
 import com.github.mikesafonov.operatorbot.service.TimetableService;
 import com.github.mikesafonov.operatorbot.service.UserService;
 import org.junit.jupiter.api.Assertions;
@@ -12,7 +13,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 
-public class CommandHandlerFactoryTest {
+public class MessageHandlerFactoryTest {
     @Mock
     private TimetableService timetableService;
     @Mock
@@ -29,37 +30,37 @@ public class CommandHandlerFactoryTest {
 
     @Test
     public void shouldReturnDefaultHandler() {
-        CommandHandler actual = commandHandlerFactory.createNewHandler(Command.NONE);
+        MessageHandler actual = commandHandlerFactory.createNewHandler(Command.NONE);
         Assertions.assertEquals(DefaultHandler.class, actual.getClass());
     }
 
     @Test
     public void shouldReturnHelpHandler() {
-        CommandHandler actual = commandHandlerFactory.createNewHandler(Command.HELP);
+        MessageHandler actual = commandHandlerFactory.createNewHandler(Command.HELP);
         Assertions.assertEquals(HelpHandler.class, actual.getClass());
     }
 
     @Test
     public void shouldReturnRoleHandler() {
-        CommandHandler actual = commandHandlerFactory.createNewHandler(Command.ROLE);
+        MessageHandler actual = commandHandlerFactory.createNewHandler(Command.ROLE);
         Assertions.assertEquals(RoleHandler.class, actual.getClass());
     }
 
     @Test
     public void shouldReturnStartHandler() {
-        CommandHandler actual = commandHandlerFactory.createNewHandler(Command.START);
+        MessageHandler actual = commandHandlerFactory.createNewHandler(Command.START);
         Assertions.assertEquals(StartHandler.class, actual.getClass());
     }
 
     @Test
     public void shouldReturnWhoHandler() {
-        CommandHandler actual = commandHandlerFactory.createNewHandler(Command.WHO);
+        MessageHandler actual = commandHandlerFactory.createNewHandler(Command.WHO);
         Assertions.assertEquals(WhoHandler.class, actual.getClass());
     }
 
     @Test
     public void shouldReturnAdminHandler() {
-        CommandHandler actual = commandHandlerFactory.createNewHandler(Command.ADD_USER);
+        MessageHandler actual = commandHandlerFactory.createNewHandler(Command.ADD_USER);
         Assertions.assertEquals(AddHandler.class, actual.getClass());
     }
 }
