+++ src/test/java/com/github/mikesafonov/operatorbot/messageReciever/handler/WhenMeHandlerTest.java
@@ -3,7 +3,7 @@ package com.github.mikesafonov.operatorbot.messageReciever.handler;
 import com.github.mikesafonov.operatorbot.command.Command;
 import com.github.mikesafonov.operatorbot.command.ParsedCommand;
 import com.github.mikesafonov.operatorbot.command.Parser;
-import com.github.mikesafonov.operatorbot.handler.internal.WhenMeHandler;
+import com.github.mikesafonov.operatorbot.handler.command.internal.WhenMeHandler;
 import com.github.mikesafonov.operatorbot.model.Status;
 import com.github.mikesafonov.operatorbot.model.Timetable;
 import com.github.mikesafonov.operatorbot.model.User;
